Very shaky way of parsing multiple fields in a struct, still need to deal with the general case

+++ examples/int_and_long/test.h
@@ -1,4 +1,4 @@
 struct foo {
     int a;
-    int big_one;
+    long big_one;
 };
